Update on the Camera - Version A
Camera follows the player after a slight delay, it makes the movement
smooth.
I plan to implement version B soon once i find the time to do so.
Version B Implementation Plan:
Set 1 Screen Size as a grid, and transition the camera only when Player
character enters the Screen Edge and trigger it to move.
It would be the Mega Man camera Implmentation where the entire Camera
Shifts top down or sides dependon on player interaction with Screen edge
triggers.